Split up method-unit-actions
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/691

The method-unit-actions file collects a lot of different operations.
These have been split up into multiple files based on semantically
coherent functionality.
Also, usage functions for the various commands have been added.
