+++ src/client/method-daemon-reload.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright Contributors to the Eclipse BlueChi project
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+#pragma once
+
+#include "libbluechi/cli/command.h"
+
+int method_daemon_reload(Command *command, void *userdata);
+void usage_method_daemon_reload();
