+++ src/client/method-reset-failed.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright Contributors to the Eclipse BlueChi project
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+#pragma once
+
+#include "libbluechi/cli/command.h"
+
+int method_reset_failed(Command *command, void *userdata);
+void usage_method_reset_failed();
